New feature (Auto Restart): index.ts & Logger.ts
Added the logic for auto-restart: If the child process does not output logs for an extended period (as defined in config.json as restartTime), it will exit abnormally. If the child process exits abnormally, the main process will attempt to fork a new child process with the same account data for a maximum of five attempts.